--- src/main/java/com/garamgaebi/GaramgaebiServer/domain/profile/dto/S3Profile.java
@@ -1,9 +1,0 @@
-package com.garamgaebi.GaramgaebiServer.domain.profile.dto;
-
-import lombok.Data;
-
-@Data
-public class S3Profile {
-    private long memberIdx;
-    private String nickName;
-}
