+++ src/main/java/com/garamgaebi/GaramgaebiServer/domain/profile/dto/reqeust/S3ProfileReq.java
@@ -1,0 +1,9 @@
+package com.garamgaebi.GaramgaebiServer.domain.profile.dto.reqeust;
+
+import lombok.Data;
+
+@Data
+public class S3ProfileReq {
+    private long memberIdx;
+    private String nickName;
+}
