+++ src/main/java/com/garamgaebi/GaramgaebiServer/domain/profile/dto/reqeust/UpdateSNSReq.java
@@ -1,4 +1,4 @@
-package com.garamgaebi.GaramgaebiServer.domain.profile.dto;
+package com.garamgaebi.GaramgaebiServer.domain.profile.dto.reqeust;
 
 
 import lombok.*;
