+++ src/main/java/com/garamgaebi/GaramgaebiServer/domain/profile/dto/response/GetCareerListRes.java
@@ -1,17 +1,17 @@
-package com.garamgaebi.GaramgaebiServer.domain.profile.dto;
+package com.garamgaebi.GaramgaebiServer.domain.profile.dto.response;
 
-
+import com.garamgaebi.GaramgaebiServer.domain.profile.entity.vo.IsWorking;
 import lombok.*;
 
 @Getter
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PUBLIC)
-public class UpdateCareerReq {
+public class GetCareerListRes {
     private Long careerIdx;
     private String company;
     private String position;
-    private String isWorking;
+    private IsWorking isWorking;
     private String startDate;
     private String endDate;
 }
